refactor: update chord-sheet schema
