+++ frontend/src/storage/types/index.ts
@@ -3,10 +3,10 @@
  * Provides clean imports for all storage-related types
  */
 
-export type { StoredChordSheet } from './chord-sheet';
-export type { SearchCacheEntry } from './search-cache';
+export type { StoredChordSheet } from "./stored-chord-sheet";
+export type { SearchCacheEntry } from "./search-cache";
 export type {
   ChordiumDBSchema,
   ChordSheetStore,
   SearchCacheStore,
-} from './schema';
+} from "./schema";
